+++ GauchoGrub/app/src/main/java/edu/ucsb/cs/cs190i/gauchogrub/gauchogrub/menu/Meal.java
@@ -1,0 +1,4 @@
+package edu.ucsb.cs.cs190i.gauchogrub.gauchogrub.menu;
+
+public class Meal extends NamedEnum {
+}
