+++ GauchoGrub/app/src/main/java/edu/ucsb/cs/cs190i/gauchogrub/gauchogrub/menu/MenuItem.java
@@ -1,0 +1,18 @@
+package edu.ucsb.cs.cs190i.gauchogrub.gauchogrub.menu;
+
+import com.google.gson.annotations.SerializedName;
+
+public class MenuItem {
+    @SerializedName("Title")
+    public String title;
+    @SerializedName("MenuCategory")
+    public MenuCategory menuCategory;
+    @SerializedName("MenuItemType")
+    public MenuItemType menuItemType;
+    @SerializedName("Id")
+    public int menuItemID;
+    @SerializedName("TotalRatings")
+    public int totalRatings;
+    @SerializedName("TotalPositiveRatings")
+    public int totalPositiveRatings;
+}
