+++ GauchoGrub/app/src/main/java/edu/ucsb/cs/cs190i/gauchogrub/gauchogrub/menu/NamedEnum.java
@@ -1,0 +1,8 @@
+package edu.ucsb.cs.cs190i.gauchogrub.gauchogrub.menu;
+
+import com.google.gson.annotations.SerializedName;
+
+public class NamedEnum {
+    @SerializedName("Name")
+    public String name;
+}
