--- GauchoGrubAndroid/app/src/main/java/com/g10/gauchogrub/menu/MenuCategory.java
@@ -1,4 +1,0 @@
-package com.g10.gauchogrub.menu;
-
-public class MenuCategory extends NamedEnum {
-}
