--- GauchoGrubAndroid/app/src/main/java/com/g10/gauchogrub/menu/NamedEnum.java
@@ -1,8 +1,0 @@
-package com.g10.gauchogrub.menu;
-
-import com.google.gson.annotations.SerializedName;
-
-public class NamedEnum {
-    @SerializedName("Name")
-    public String name;
-}
