--- GauchoGrubAndroid/app/src/main/java/com/g10/gauchogrub/utils/gson/DateTimeDeserializer.java
@@ -1,18 +1,0 @@
-package com.g10.gauchogrub.utils.gson;
-
-import com.google.gson.JsonDeserializationContext;
-import com.google.gson.JsonDeserializer;
-import com.google.gson.JsonElement;
-import com.google.gson.JsonParseException;
-
-import org.joda.time.DateTime;
-import org.joda.time.format.ISODateTimeFormat;
-
-import java.lang.reflect.Type;
-
-public class DateTimeDeserializer implements JsonDeserializer<DateTime> {
-    @Override
-    public DateTime deserialize(JsonElement json, Type type, JsonDeserializationContext context) throws JsonParseException {
-        return ISODateTimeFormat.dateTimeParser().parseDateTime(json.getAsString());
-    }
-}
